Build a vector from the origin approximating the minimum and maximum values

Build a vector between the minimum and maximum ensuring that the given origin value is part of the returned vector.

Usage

grid_helper(o, min, max, res)

Arguments

o

a numeric(1). Origin.

min

a numeric(1). Mininum value.

max

a numeric(1). Maximum value.

res

a numeric(1). Resultuion.

Value

A numeric.